refactor: organize front-end code
