+++ .gitignore
@@ -187,3 +187,4 @@ fabric.properties
 # idea folder, uncomment if you don't need it
 .idea
 /db/.env
+!front-end/src/lib/
