+++ front-end/AGENTS.md
@@ -1,0 +1,9 @@
+# Front-end Guidelines
+
+This directory contains the React dashboard built with Vite.
+
+- Place small reusable components in `src/components/`.
+- Larger pages should live under `src/pages/`.
+- Shared utilities belong in `src/lib/`.
+- Update import paths to match this structure when adding new code.
+- Before submitting a pull request run `npm install` and `npm run build` to ensure the build succeeds.
