+++ front-end/src/components/ClanModal.jsx
