+++ front-end/src/components/MobileTabs.jsx
