+++ front-end/src/components/RiskBadge.jsx
