+++ front-end/src/components/RiskRing.jsx
